+++ CHANGELOG.md
@@ -26,7 +26,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Update HPO terms using downloaded definitions files
 - Option to use alternative flask config when running `scout serve`
 - Requirement to use loqusdb >= 2.5 if integrated
-- Keep variant tagged status (dismissed, manual rank, mosaic) upon variant re-upload
+- Keep variant actions (dismissed, manual rank, mosaic, acmg, comments) upon variant re-upload
 
 ### Fixed
 - Do not display Pedigree panel in cancer view